Separate core NAM code from other plugin dsp code (#21)
This PR separates the core NAM code from other plugin dsp code and removes IPlug dependencies for NAM code.

I tried to keep this as simple and targeted as possible. It looks like bigger changes than it is because of the separation of dsp.h/dsp.cpp files into an IPlug dsp base class for noise gate, IR and EQ and the core dsp code for the network models.

I kept the input to the model processing as "double" for now - since that is what the IPlug plugin is using. I also haven't made any changes to the redundant gain handling - I'll save that for a separate PR.

With the changes in this PR, it is now easy to use the core NAM code directly in a context other than the IPlug plugin.


Squash of the following commits:

* Separate core NAM code from other plugin dsp code and remove IPlug dependency for NAM code.

* Move cncp.cpp/.h to NAM subdir

* Merge sdatkinson:main into mikeoliphant:separate_nam

---------

Co-authored-by: Mike Oliphant <oliphant@nostatic.org>